requirements.txt
docx==0.2.4
docxtpl==0.15.2
face_recognition==1.3.0
gspread==5.1.1
imutils==0.5.4
matplotlib==3.4.1
numpy==1.19.5
onnx==1.10.1
opencv_python_headless==4.5.4.60
pandas==1.2.3
Pillow==9.0.0
pydantic==1.8.1
pygame==2.1.2
pyserial==3.5
python_docx==0.8.11
requests==2.26.0
scikit_learn==1.0.2
streamlit==1.5.0
tensorflow==2.7.0
tensorflow_hub==0.12.0
tf2onnx==1.9.3
tf_nightly==2.9.0.dev20220121
torch==1.10.0
torchvision==0.11.1
xlwings==0.25.3
